MLservice/requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
